Initial implementation of OpenPBR material system.
This removes the old material system and replaces it with a new system
inspired by OpenPBR. The GLTF importer is rewritten to use this system.

The implementation is very slow at the moment, but it's higher quality
and much more feature complete, including transmission, fuzz and clear coat.

It is tested against the reference implementation of OpenPBR, meaning
this iteration is very likely to be correct, and will be used to test
future versions against for correctness.